+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
 ### Fixed
 - Fixed a bug where submitting a blank line item quantity throws "Unsupported operand types: int + string" error in PHP 8. ([#2125](https://github.com/craftcms/commerce/issues/2125))
+- Fixed a bug where updating a state on an address on the Order edit page wouldn’t persist. ([#2136](https://github.com/craftcms/commerce/issues/2136))
 
 ##  3.3.1.1 - 2021-05-09